.zshrc: export COLORTERM=truecolor in WSL
Windows Terminal has full color support, but
programs don't know about it because this whole
part of the TTY/PTY ecosystem is bad (the wrong
parts of the stack are forced to know/provide
the necessary capability information).

This was mostly just a mild nuisance (Emacs
in the terminal would use blatantly wrong
colors), but I just noticed degrated color
distinctiveness in my Git Delta diffs because
of it (and worse, wasted my time tuning my
colors, and worst, gaslit myself by supposing
that either my vision was changing or I had
overlooked this legibility suboptimality in
my color choices), and that's the last straw.