.gitconfig: delta: revert primary moved-to color
This reverts the cyan color change in commit
5adf1aa0a0faa5798765dce4314243820ed88410.

(The change to the cyan was last-minute, less
confident/obvious/strict of a win than the
others. The dark purple was darker and a bit
ambiguous in bad lighting, both vs my red and
vs my black and dark gray backgrounds. The
dark cyan was never really ambiguous, I just
wanted to try optimizing its distinctiveness
versus the newly darkened green.)

The downsides of pushing it towards blue were

1. it becomes a little more ambiguous with
   the two purples, especially when adjacent
   to them and viewed in bright sunlight; and

2. less similarity with the other cyan (and
   thus less intuitive interpretability).