Emacs: datetime-parse: refactor (deduplicate)
Debatable if this is better or worse.

The gain is it DRYs the bulk of the repetition
from the main parse loop into a single macro,
explicates that the year is excluded from
ambiguous integer binding, makes it unnecessary
to remember to pair the assignments that should
be paired, and improves signal-to-noise within
the source of the loop body's cases.

The cost is pushing the non-local
entangelements one layer deeper in the
stack (macro expansions referring to
variable names that are not passed in),
and entangeling even more (changing how
the value is handled based on if being
passed into the macro as a specific
source-level variable name).