Emacs: occasional-progress-counter
Application of `occasional` for printing
a progress indicator in the form of a
"done/total" counter.

(`progress-reporter` only provides percentage
and spinner options, requires the caller to
track the count, and percentage needs to know
the total in advance. A counter gives a more
precise view into what's happening, gives a
more accurate/concrete sense of speed, and
does so even for streaming operations which
don't know the total.)