Emacs: occasional: run code "occasionally"
Provide support for doing something every
once in a while -- for example, echoing a
progress message or updating the screen in
the middle of a long command.

"Every once in a while" defaults to every
10th of a second (plus usual delay noise).

Versus `progress-reporter`, this code is a
general/flexible base for building things
such as `progress-reporter`.