Emacs: evil-dired & music: show progress if slow
This comes up when manipulating hundreds or
thousands of files at once -- for example,
organizing years of accumulated photos. Or
when loading my entire music library at once
to just play everything.

The worst part of the freeze in those cases
is not having any feel for how long it will
take or that progress is being made. Seeing
a "done/total" count provides a good feel of
that.

The UX I actually want is non-blocking, with
live incremental updates and an initial visual
confirmation that resembles the real thing.
For example, pasted entries could be inserted
immediately as placeholders with a distinct
face, each turning to normal as the actual
underlying operation finishes. (Deletion could
delete immediately. For files, keep updating
the pasticle in the background, and if it's
pasted before trashing finishes, maybe abort
the trashing and just do the moves directly.)
But for now, a counter is good enough.