Emacs: music: optimize playlist manipulations
Manipulating hundreds of songs at once is now
reliably subsecond, and often feels instant.
And that's on a Pixel 5.

The one blemish is that if a replacing paste
is ever big enough to trigger the progress
counter, it will show two rounds of "mpv
commands" and "mpv replies" instead of one.
We could in principle combine them, but in
practice I don't expect to ever see the
progress counter anymore, so the refactor
time and effort isn't worth it (replacing
paste would have to reimplement adding and
grow its own specialized undo function).